fix(review): absent-route status, DSN escapes, int offsets, API docs
Seven review findings on this PR. The sharpest one is last.

splitKeywordValueDSN rejected a DSN that works. lib/pq's unquoted-value
loop consumes a backslash-escaped byte (connector.go:759-771), so
`password=a\ b` is one value to the driver and `key "b" has no value` from
us. The unquoted branch now consumes escaped pairs. A trailing lone
backslash stays a literal here: we re-emit raw text, so the driver's own
parse is authoritative, and erroring to match lib/pq exactly would reject
DSNs C libpq accepts — the same defect class as the finding itself.

The connect fixture interpolated the password unquoted. An empty one, which
is what trust and peer auth give, produced `password= dbname=...`; both
libpq and our tokenizer skip whitespace after the equals sign, so dbname
became the password value and the test failed blaming the rewrite instead
of the fixture. Quoted now, with backslash and quote escaping so a hostile
password does not swap one fixture bug for another.

buildBacktickRuns narrowed offsets to int32 while Render has no size limit,
so a block over MaxInt32 could wrap and select an invalid slice bound. Now
int. Measured: +17% B/op on the adversarial 1 MB shape, unchanged at 424
B/op elsewhere, ns/op in noise. The 2 GiB proof is env-gated
(MARKDOWN_BIGINPUT=1) because 6.2 GiB peak RSS would OOM a CI runner —
same precedent as PGTEST_REQUIRED.

The carousel failure messages passed two values to a three-slot format,
dropping the pre-nav baseline — the reading that separates a real leak from
a probe that never armed.

llmmd's limit assertion only rejected "max 100", so any OTHER wrong cap
passed silently.

And the one that mattered most: both read-only "reality checks" only
rejected success codes. That block is the premise of everything after it —
the doc may omit those routes because the server does not serve them — and
rejecting only 200 does not establish that. Mounting the view with writes
proves it: POST answers 415, which the old check accepted silently. Both
sites now require 404 or 405, and with writes mounted
TestView_LLMMDMatchesReadOnlyMount fails as it always should have.

Docs ship with the change per the repo rule: entity-declarations.md and
migrations.md cover the new mount-aware llm.md, and the CHANGELOG entry
sits under Unreleased because the predicate signature is breaking for
direct callers.

+++ core/markdown/inline.go
@@ -210,17 +210,20 @@ func findClosingDelim(runs *backtickRuns, s string, start int, delim byte, run i
 type backtickRuns struct {
 	// start/lens list each maximal run's start offset and full length, in
 	// document order. byLen maps a run length to the ascending starts of
-	// all runs with that length.
-	start []int32
-	lens  []int32
-	byLen map[int32][]int32
+	// all runs with that length. All three are int: Render caps no
+	// document size, so an int32 offset wraps on a block over
+	// math.MaxInt32 — a negative "open" reaches the caller and slices
+	// out of range (pinned by TestMarkdown_GiantBacktickRunNoInt32Wrap).
+	start []int
+	lens  []int
+	byLen map[int][]int
 }
 
 // buildBacktickRuns makes the index for s. Callers only build it when s
 // contains a backtick (strings.IndexByte gate), keeping backtick-free
 // blocks on the fast path.
 func buildBacktickRuns(s string) *backtickRuns {
-	b := &backtickRuns{byLen: make(map[int32][]int32)}
+	b := &backtickRuns{byLen: make(map[int][]int)}
 	for i := 0; i < len(s); {
 		if s[i] != '`' {
 			i++
@@ -230,10 +233,10 @@ func buildBacktickRuns(s string) *backtickRuns {
 		for j < len(s) && s[j] == '`' {
 			j++
 		}
-		l := int32(j - i)
-		b.start = append(b.start, int32(i))
+		l := j - i
+		b.start = append(b.start, i)
 		b.lens = append(b.lens, l)
-		b.byLen[l] = append(b.byLen[l], int32(i))
+		b.byLen[l] = append(b.byLen[l], i)
 		i = j
 	}
 	return b
@@ -246,15 +249,15 @@ func buildBacktickRuns(s string) *backtickRuns {
 // a backtick of a block this index was built for; mid-run positions use
 // the remaining suffix as the opener, the same rule a linear scan had.
 func (b *backtickRuns) findCodeEnd(i int) (end, open int) {
-	r := sort.Search(len(b.start), func(k int) bool { return b.start[k] > int32(i) }) - 1
+	r := sort.Search(len(b.start), func(k int) bool { return b.start[k] > i }) - 1
 	// Callers only query at backticks, so run r contains i.
-	open = int(b.start[r]+b.lens[r]) - i
-	starts := b.byLen[int32(open)]
-	k := sort.Search(len(starts), func(k int) bool { return starts[k] >= int32(i+open) })
+	open = b.start[r] + b.lens[r] - i
+	starts := b.byLen[open]
+	k := sort.Search(len(starts), func(k int) bool { return starts[k] >= i+open })
 	if k == len(starts) {
 		return -1, open
 	}
-	return int(starts[k]), open
+	return starts[k], open
 }
 
 // parseLink parses [text](url) starting at the '[' at position i.
